Initial commit for this small utility
Given I often need to run a same Git command over a bunch of
subdirectories, I developed this small Git addon as an helper
for this task.

I used it since one year for now, did a small refactoring on it,
and it’s time now to issue it and properly version-control it.